Module_5/taxi-pipeline/pipeline/assets/ingestion/requirements.txt
# Python dependencies for trips.py ingestion asset.
# Bruin finds this file next to the asset and installs before running.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
pandas>=2.0.0
pyarrow>=14.0.0
python-dateutil>=2.8.0
tzdata>=2024.1 # required for datetime.now() on Windows
